### 2020-12-10 #2882 @usethedata
Sorry. This week blew up on me with the American Geophysical Union meeting and a family member ER trip.  It will be the weekend before I can start on this.
