### 2021-01-12 #2882 @usethedata
@kyleniemeyer -- Per the discussion with the editors in chief, I believe the intent was to assign someone else to edit this paper, given the challenges in my world.

@ctodd26labs -- My apologies for the lack of action here.  For a variety of personal and job-related reasons, I've not had the bandwidth to work on JOSS things and am stepping away from being a topic editor.
