### 2021-01-25 #2882 @arfon
@ctodd26labs - apologies for the delay here in getting this submission moving.

Could you take a look a [this list of potential reviewers](https://bit.ly/joss-reviewers) and identify a few people who would be good candidates to review this submission? (Please list their GitHub handles without a `@` when listing them here so as avoid accidentally pinging a bunch of people).
