### 2021-01-26 #2882 @arfon
👋 @nidhog @fabianomenegidio @dvysardana - would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The submission under consideration is a Python package: [_Magicwand: A platform to provide high-quality, reliable, and reproducible data sets for low-and-slow DDoS attacks_](https://github.com/twosixlabs/magicwand-datatool).
