### 2021-02-03 7909115
using full name for docker images (not tagging them after pulling) so that the data tool will pull them automatically if they aren't already pulled locally
