### 2021-02-03 #2882 @fabianomenegidio
Sorry. I'm away assisting in the assembly of the genome of the
Brazilian strain of SARS-CoV-2.

I will not be able to revise the article.

Em qua, 3 de fev de 2021 13:40, Christopher H. Todd <
notifications@github.com> escreveu:

> Hello,
>
> Wanted to follow-up to see if there has been any movement with this?
>
> Thank you in advance and have a good day.
>
