### 2021-02-03 69bf8d8
storing name of cic converter docker image in a var instead of hardcoding it in multiple places
