+++ magicwand-data-generator/magicwand/magicwand_config/config.py
@@ -17,5 +17,4 @@ class Config:
     Base Config Class for MAGICWAND. Will hold constants and configuration values
     """
 
-
     VERSION = "1.0.2"
