### 2021-02-10 3272c20
Update pull_images.sh
