### 2021-02-11 #3032 @whedon
[REVIEW]: Magicwand: A platform to provide high-quality, reliable, and reproducible data sets for low-and-slow DDoS attacks
**Submitting author:** @ctodd26labs (<a href="http://orcid.org/0000-0003-1453-265X">Christopher Todd</a>)
**Repository:** <a href="https://github.com/twosixlabs/magicwand-datatool" target ="_blank">https://github.com/twosixlabs/magicwand-datatool</a>
**Version:** v1.0.0
