+++ magicwand-data-generator/README.md
@@ -1,4 +1,4 @@
 # Magicwand Docker Images
 
 This directory contains the code for the magicwand CLI.  
-To learn more about the CLI please visit our documentation at INSERT_URL
+To learn more about the CLI please visit our documentation at https://magicwand-datatool.readthedocs.io/en/latest/
