### 2021-03-08 #2882 @ctodd26labs
Hello @arfon,

I got the update today from this thread and I have not seen any updates on https://github.com/openjournals/joss-reviews/issues/3032 in ~16 days. Is there anything needed from me or are we choosing a new reviewer for the paper?

I appreciate you keeping an eye on it and updating the ticket.
