### 2021-03-08 #2882 @arfon
> I got the update today from this thread and I have not seen any updates on #3032 in ~16 days. Is there anything needed from me or are we choosing a new reviewer for the paper?

Let's sit tight for now. My comment earlier today was to try and find a second reviewer to go along with the one we already have (JOSS has a minimum of two reviewers per submission)
