### 2021-04-30 #3032 @arfon
@gradvohl - Please read the "Reviewer instructions &amp; questions" in the first comment above.

You have a checklist at the top of this thread (in that first comment) with the JOSS requirements. As you go over the submission, please check any items that you feel have been satisfied. There are also links to the JOSS reviewer guidelines.

Finally, just as a heads up, our bot @whedon will send an automatic reminder at some point in the next week or so but I realize you might not be able to complete your review for a number of weeks.
