### 2021-05-01 #3032 @gradvohl
# Reviewer report for JOSS paper #3032
**Paper title**: Magicwand: A platform to provide high-quality, reliable, and reproducible data sets for low-and-slow DDoS attacks.

## Review criteria
* A list of the authors of the software: :heavy_check_mark:
* Author affiliations: :wave: (sort of -- Please see the following **Author affiliations section** in the **Reviewer Comments** section)
* A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience: :heavy_check_mark:
* A clear statement of need that illustrates the purpose of the software: :heavy_check_mark:
* Mentions (if applicable) of any ongoing research projects using the software or recent scholarly publications enabled by it: :heavy_check_mark:
* A list of key references including a link to the software archive:  :wave: (sort of -- Please see the **References in the paper** section in the **Reviewer Comments** section)

## Review items
* Software license: :thumbsup: (good - MIT License)

* Documentation:
    * A statement of need: :thumbsup: (good)
    * Installation instructions: :thumbsup: (good)
    * Example usage: :thumbsup: (good)
    * API documentation: :thumbsup: (good)
    * Community guidelines: :thumbsup: (good)
    * Functionality: :thumbsup: (good)
    * Tests: :thumbsup: (good)

* Software paper
    * Authors: :heavy_check_mark:
    * A statement of need: :heavy_check_mark:
    * References: :wave: (sort of -- Please see the **References in the paper** section in the **Reviewer Comments** section)

## Test environment
The reviewer conducted the tests for this software using an AWS instance (t2.xlarge) with Amazon Linux 2.

## Reviewer comments
The following comments refer to items I considered  that could be improved a little more, although they currently met the requirements for JOSS submission.

### Authors affiliation
Regarding the authors' affiliation, I noticed that the author Anthony C. Wong is not affiliated with any organization. However, perhaps it is better to write that he is an "independent developer" or a freelancer rather than an unaffiliated. It is merely a matter of words.

### References in the Paper
Concerning the references in the article, it would be interesting that the cited scientific paper (Sharafaldin et al. 2017) had its DOI informed ([10.5220/0006639801080116](https://doi.org/10.5220/0006639801080116)). Besides, the other references (Krebs 2016, Makrushin 2013, Markoff 2008, and Plante 2015) could have links to the sites where the reader can find them.

### Suggestions on the README file in the repository
Following there are some suggestions to improve the documentation on the repository main page.
1. Just before the "Contributing to Magicwand" section, the authors said that "_For complete documentation on Magicwand, a gallery of available attacks, the configuration guide, tutorials and teaching resources, frequently asked questions, and more, please visit our documentation._" However, when I went to the docs looking for a gallery, I didn't find it with this name, neither the tutorials, teaching resources, or FAQs. I suppose the authors spread those sections through the documentation with other words such as "Attacks," "Data," and so on. Therefore, I suggest that authors create those sections in the documentation or rewrite the statement regarding the complete documentation on Magicwand.
2. For hosting the MagicWand Datasets, I suggest the authors use [Zenodo -- https://zenodo.org](https://zenodo.org). Beyond free hosting for datasets, the Zenodo allows the creation of a DOI for the datasets and the inclusion of metadata describing the data. Then other researchers in the Computer Security fields can use the data you produced.
3. Some tests did not work using a regular user. However, using ``sudo``, all the tests ran fine. Perhaps the authors can mention this information in the documentation.
4. Don't forget to replace the ``TODO_INSERT_GITHUB_URL`` with the proper URL for the repository in the ``CONTRIBUTING.md`` file.

## Conclusion
Analyzing all the aspects mentioned earlier, I rank the paper as (very) **minor revisions** needed. Most of those revisions are very easy to implement.

By the way, I would like to congratulate the authors for producing very well-documented software. I think it will be helpful in many works in the Computer Security field.
