### 2021-05-03 998a9b6
correcting bib and updating with missing references
