+++ paper/paper.bib
@@ -1,18 +1,46 @@
-@misc{krebs,
-    author    = "Brian Krebs",
-    title     = "KrebsOnSecurity Hit With Record DDoS",
-    year = {2016},
-    url       = "https://krebsonsecurity.com/2016/09/krebsonsecurity-hit-with-record-ddos/"
+@article{sharafaldin:2018,
+  author = {Sharafaldin, I., Lashkari, A. H., and Ghorbani, A. A.},
+  year = {2018},
+  title = {Toward generating a new intrusion detection dataset and intrusion traffic characterization},
+  journal = {SciTePress},
+  url = {https://www.scitepress.org/Link.aspx?doi=10.5220/0006639801080116}
 }
 
-@misc{makrushin,
-    author    = "Denis Makrushin",
-    title     = "The cost of launching a DDoS attack",
-    year = {2013},
-    url       = "https://securelist.com/the-cost-of-launching-a-ddos-attack/77784/"
+@misc{krebs:2016,
+  author    = "Brian Krebs",
+  title     = "KrebsOnSecurity Hit With Record DDoS",
+  year = {2016},
+  url       = "https://krebsonsecurity.com/2016/09/krebsonsecurity-hit-with-record-ddos/"
 }
 
-@misc{locust,
+@misc{makrushin:2013,
+  author    = "Denis Makrushin",
+  title     = "The cost of launching a DDoS attack",
+  year = {2013},
+  url       = "https://securelist.com/the-cost-of-launching-a-ddos-attack/77784/"
+}
+
+@article{russia:2008,
+  author = {Markoff, John},
+  year = {2008},
+  month = {August},
+  day = {12},
+  title = {Before the Gunfire, Cyberattacks},
+  journaltitle = {The New York Times},
+  url = {https://www.nytimes.com/2008/08/13/technology/13cyber.html}
+}
+
+@article{dota:2015,
+  author = {Chris Plante},
+  editor = {{theverge}},
+  title = {Valve's \$18 million Dota 2 tournament delayed by DDoS attack},
+  url = {https://www.theverge.com/2015/8/4/9097597/the-international-dota-2-ddos-attack-valve},
+  year = {2015},
+  lastchecked = {5/8/2019},
+  originalyear = {8/4/2015}
+}
+
+@misc{locust:2019,
   author = {Carl Byström and Jonatan Heyman and Joakim Hamrén and Hugo Heyman},
   title = {Locust},
   year = {2019},
@@ -21,7 +49,7 @@
   howpublished = {\url{https://github.com/locustio/locust}}
 }
 
-@misc{goloris,
+@misc{goloris:2014,
   author = {Aliaksandr Valialkin},
   title = {Goloris},
   year = {2014},
@@ -30,7 +58,7 @@
   howpublished = {\url{https://github.com/valyala/goloris}}
 }
 
-@misc{sht,
+@misc{sht:2011,
   author = {Sergey Shekyan},
   title = {SlowHTTPTest},
   year = {2011},
@@ -39,7 +67,7 @@
   howpublished = {\url{https://github.com/shekyan/slowhttptest}}
 }
 
-@misc{killapache,
+@misc{killapache:2011,
   author = {Miroslav Stampar},
   title = {KillApachePy},
   year = {2011},
@@ -47,31 +75,11 @@
   journal = {GitHub repository},
   howpublished = {\url{https://github.com/tkisason/KillApachePy/}}
 }
-@misc{sockstress,
+@misc{sockstress:2012,
   author = {Taylor Hornby},
   title = {Sockstress},
   year = {2012},
   publisher = {GitHub},
   journal = {GitHub repository},
   howpublished = {\url{https://github.com/defuse/sockstress}}
-}
-
[…]
