+++ paper/paper.md
@@ -24,7 +24,7 @@ affiliations:
    index: 1
  - name: Unaffiliated
    index: 2
-date: xx September 2020
+date: xx May 2021
 bibliography: paper.bib
 ---
 
