### 2021-05-03 bf12d15
updating citations to conform to JOSS standard. removing from .md and letting auto-insert from .bib (adds more details as well and verifies all are used)
