+++ paper/paper.md
@@ -34,9 +34,9 @@ bibliography: paper.bib
 
 # Statement of Need
 
-Distributed denial-of-service (DDoS) attacks remain a pervasive cybersecurity threat. While various commercial services such as Akamai and CloudFlare are effective in mitigating traditional, high-volume DDoS attacks, LSDDoS attacks present a different kind of threat. These attacks stealthily degrade server performance through cleverly crafted transmissions of data, rather than brute-force floods of traffic. The adversary sends low-bandwidth requests that slowly transmit data and consume a disproportionate amount of the server's resources. The relatively small amount of traffic needed to cause a denial-of-service has two important consequences: these attacks are both easier to launch and more difficult to detect, in comparison to high-volume DDoS.
+Distributed denial-of-service (DDoS) attacks remain a pervasive cybersecurity threat [@krebs:2016], [@dota:2015], and [@russia:2008]. While various commercial services such as Akamai and CloudFlare are effective in mitigating traditional, high-volume DDoS attacks, LSDDoS attacks present a different kind of threat [@makrushin:2013]. These attacks stealthily degrade server performance through cleverly crafted transmissions of data, rather than brute-force floods of traffic. The adversary sends low-bandwidth requests that slowly transmit data and consume a disproportionate amount of the server's resources. The relatively small amount of traffic needed to cause a denial-of-service has two important consequences: these attacks are both easier to launch and more difficult to detect, in comparison to high-volume DDoS.
 
-In order to create a robust LSDDoS defense, researchers in the security community must have access to sufficient data capturing how these attacks behave. The most widely-adopted public source of data for LSDDoS was generated by the Canadian Institute for Cybersecurity (CIC) at the University of New Brunswick (Sharafaldin, 2017). However, the vast majority of instances in this dataset are from high-volume attacks, and the dataset does not contain any benign traffic directed towards the victim host. Additionally, the range of LSDDoS attacks employed is limited. In an effort to facilitate the generation of more thorough LSDDoS datasets, we are introducing our traffic generation tool `Magicwand`.
+In order to create a robust LSDDoS defense, researchers in the security community must have access to sufficient data capturing how these attacks behave. The most widely-adopted public source of data for LSDDoS was generated by the Canadian Institute for Cybersecurity (CIC) at the University of New Brunswick [@sharafaldin:2018]. However, the vast majority of instances in this dataset are from high-volume attacks, and the dataset does not contain any benign traffic directed towards the victim host. Additionally, the range of LSDDoS attacks employed is limited. In an effort to facilitate the generation of more thorough LSDDoS datasets, we are introducing our traffic generation tool `Magicwand`.
 
 Through its development, these key goals informed the design and implementation:
 
@@ -53,11 +53,11 @@ Through its development, these key goals informed the design and implementation:
 
 ## Attack Clients
 
-The set of implemented attack images consists of Slowloris (Valialkin, 2014), R U Dead Yet (Shekyan, 2011), Slow Read (Shekyan, 2011), Apache Killer (Stampar, 2011), Sockstress (Hornby, 2012), and a second implementation of Slowloris (Shekyan, 2011).  The JSON files provide configurations that can scale the intensity of attack and adjust other attack-specific-specific parameters (e.g. threads per connection for Apache Killer).
[…]
