### 2021-05-03 f683ed5
Note about sudo (Linux Docker (and other cses) can need root access depending on permissions setup)
