### 2021-05-03 #3032 @ctodd26labs
Hello @gradvohl,

Thank you for going through everything and reviewing the paper/repo. Your comments/information is incredibly helpful.

With that said, I have created https://github.com/twosixlabs/magicwand-datatool/pull/15, which addresses the comments you have made with 2 exceptions:

1. My team is talking internally about posting data to Zenodo (and anywhere else). We wanted to get the paper/repo approved before publishing anything in case we needed to make code changes that could have impacted what we published. There are also ongoing conversations about what exact dataset we plan to publish (many options), so things to work out there.
2. Anthony Wong's affiliation is specific from our contract/approval with the government (he worked for Two Six and left during the publication). Is this ok to keep (will be harder to change on our end and need approvals)
