### 2021-05-03 #3032 @gradvohl
Hello @ctodd26labs,
Concerning your pull request, I think it is almost done. There are two things I would like to clarify.
1. Regarding the **Note** about running as ``sudo`` (``README.md``, line 44), is it really necessary to have ``bash`` twice? Wouldn't `sudo scripts/pull_images.sh` enough?
2. Looking at the ``paper.bib``, in ``sharafaldin:2018`` reference, instead of the ``URL`` field, it's better that you include the DOI field like this: ``doi = {10.5220/0006639801080116},``
These were the only things missing. I believe the rest is ok.
Best.

P.S. I am not sure if I can tag the modifications as _latest_. Perhaps @arfon can help us with that.
