### 2021-05-04 #3032 @ctodd26labs
My apologies, my request was to approve the PR once it looks good (which I have changes to make based on your comment).

Once you approve, I will merge and tag the repo as v1.0.4 (currently tagged at v1.0.3)
