### 2021-05-09 #3032 @nidhog
@ctodd26labs all I can add to this is:
* References in the paper do not seem to be cited in the text (unless I missed something, the reference list is not referenced from the paper's text)
* While there is some reasonable documentation there is no detailed real-world example that shows how magic-wand is used.

Other than that LGTM
