### 2021-05-10 #3032 @ctodd26labs
Thanks @nidhog for going through the paper as well.

For the references, they are built automatically from the .bib file (added into the .pdf). We followed https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography as the example and https://github.com/twosixlabs/magicwand-datatool/blob/joss_revisions/paper/paper.pdf seems to have the references correct. Please confirm.

For the real-world example, we describe some use cases and existing datasets in the "statement of need" section that the magicwand tool creates similar (but configurable) output for. The application docs have a walkthrough of how to generate a dataset that is calibrated to a user's system. Can you advise a bit more about what you think we should be putting in the paper?

We will adjust accordingly; thanks again for the input.
