### 2021-05-13 #3032 @nidhog
@ctodd26labs great, yes that revision does have the references properly included.

And for sure, I think the described use cases and documentation are good enough and sufficient for someone to get started. It just seemed like it would help in the documentation (not in the paper itself) to have an example tutorial, by choosing one of these examples and going through all the steps needed. For example, going through the Apache Killer scenario, and all the steps involved to run the components, adjust threads etc.
But just to be clear I think that the paper does not need this adjustment at all it all looks good to me, this is just a suggestion to make future interested people more likely to use magicwand. ;)
