### 2021-05-13 #3032 @arfon
@ctodd26labs - I think the DOI suggested by Whedon (`10.5220/0006639801080116`) is correct. Could you please add `doi=10.5220/0006639801080116` to that BibTeX entry?
