### 2021-05-19 4e8fdb7
adding message about venvs in README as the best way to install magicwand
