### 2021-05-19 14a2943
loosening dependencies for security patches to allow updates for pandas and similar libs which are necessary to be compatible with python 3.9
