### 2021-05-19 7dd91a4
adding missing doi for journal entry
