+++ paper/paper.bib
@@ -3,7 +3,8 @@
   year = {2018},
   title = {Toward generating a new intrusion detection dataset and intrusion traffic characterization},
   journal = {SciTePress},
-  url = {https://www.scitepress.org/Link.aspx?doi=10.5220/0006639801080116}
+  url = {https://www.scitepress.org/Link.aspx?doi=10.5220/0006639801080116},
+  doi = {10.5220/0006639801080116}
 }
 
 @misc{krebs:2016,
