### 2021-05-19 #3032 @ctodd26labs
Hello @arfon,

Sorry for the delay everyone. Things have been super busy and I've been trying to get to this

1. 10.5281/zenodo.4774258, including example data, has been published (https://zenodo.org/record/4774258#.YKV-1pNKjmE). Please let me know if that looks good.
2. updated the DOI suggested by Whedon
3. 10.5281/zenodo.4774301, includes the tagged software with all of the requested changes (https://zenodo.org/record/4774301#.YKWEN5NKhhE). Please let me know if that is what was requested
4. https://github.com/twosixlabs/magicwand-datatool/releases/tag/1.0.4 is the latest release after all of the changes have been submitted
5. https://pypi.org/project/magicwand/ has been published with the updated docs.

Again, thank you all for the help and reviews. Please let me know if there is anything I am missing or need to update with this current submission
